1. Unit tests for FreeRTOS wrapper under Arduino mode ready 2. Sensor spoof ready

+++ gyros/Arduino code/spoof_sensor/archive/slave_receiver/slave_receiver.ino
@@ -27,7 +27,7 @@ void loop()
 // this function is registered as an event, see setup()
 void receiveEvent(int howMany)
 {
-  char data[sizeof(float)];
+  char data[sizeof(double)];
   float y;
   int ii = 0;
   while(Wire.available()) // loop through all but the last
